refactor: Update sdk package when opening the project instead of when loading the projects
