+++ packages/renderer/src/modules/store/analytics/track.ts
@@ -10,7 +10,7 @@ trackAction(workspaceActions.createProject.fulfilled, 'Create Project', async ac
   cols: action.payload.layout.cols,
 }));
 
-trackAction(editorActions.setProject, 'Open Project', async action => ({
+trackAction(editorActions.setProject.fulfilled, 'Open Project', async action => ({
   project_id: action.payload.id,
   project_name: action.payload.title,
 }));
